+++ plugins/experimental/jax_fingerprint/CMakeLists.txt
@@ -22,13 +22,13 @@ add_atsplugin(
   userarg.cc
   header.cc
   log.cc
-  ja3/ja3_method.cc
-  ja3/ja3_utils.cc
-  ja4/ja4_method.cc
+  ja3/method.cc
+  ja3/utils.cc
+  ja4/method.cc
   ja4/ja4.cc
   ja4/datasource.cc
   ja4/tls_client_hello_summary.cc
-  ja4h/ja4h_method.cc
+  ja4h/method.cc
   ja4h/ja4h.cc
   ja4h/datasource.cc
   ja4_common/utils.cc
@@ -41,12 +41,12 @@ verify_remap_plugin(jax_fingerprint)
 if(BUILD_TESTING)
   add_executable(
     test_jax
-    ja3/test_ja3.cc
-    ja3/ja3_utils.cc
-    ja4/test_ja4.cc
+    ja3/test.cc
+    ja3/utils.cc
+    ja4/test.cc
     ja4/ja4.cc
     ja4/datasource.cc
-    ja4h/test_ja4h.cc
+    ja4h/test.cc
     ja4h/ja4h.cc
     ja4h/datasource.cc
     ja4_common/utils.cc
