+++ plugins/experimental/jax_fingerprint/ja3/test.cc
@@ -22,7 +22,7 @@
 
  */
 
-#include "ja3_utils.h"
+#include "utils.h"
 
 #include <catch2/catch_test_macros.hpp>
