+++ plugins/experimental/jax_fingerprint/ja4h/method.cc
@@ -24,7 +24,7 @@
 #include <plugin.h>
 #include <context.h>
 #include <ja4_common/utils.h>
-#include "ja4h_method.h"
+#include "method.h"
 #include "ja4h.h"
 #include "datasource.h"
